Fix use of nanoid
`nanoid` was erroneously flagged as an external dependency in Rollup,
however, it was not listed as a `peerDependency` in package.json.

This commit fixes this by removing the flag in the Rollup configuration.
This required adding the `@rollup/plugin-node-resolve` dependency since
this is the first actual dependency that needs to be bundled with the
library.

This commit also updates `nanoid` to the most recent version. This
version is tree-shakeable and it allows using the non-secure version
which is very, very tiny.